+++ weblate/trans/models/conf.py
@@ -209,5 +209,7 @@ class WeblateConf(AppConf):
     # Following probably should not be configured
     COMPONENT_NAME_LENGTH = 100
 
+    SUGGESTION_CLEANUP_DAYS = None
+
     class Meta(object):
         prefix = ''